Применяет модульность
