+++ index.html
@@ -229,6 +229,6 @@
     </div>
   </template>
 
- <script src="js/main.js"> </script>
+ <script type="module" src="js/main.js"></script>
 </body>
 </html>
